+++ lib/src/client.dart
@@ -69,6 +69,12 @@ class Client extends http.BaseClient {
   /// Callback to be invoked whenever the credentials refreshed.
   final CredentialsRefreshedCallback? _onCredentialsRefreshed;
 
+  /// Callback to be invoked whenever the credentials started to refresh.
+  final CredentialsRefreshingCallback? _onCredentialsRefreshing;
+
+  /// Callback to be invoked whenever the credentials failed to refresh.
+  final CredentialsRefreshFailedCallback? _onCredentialsRefreshFailed;
+
   /// Whether to use HTTP Basic authentication for authorizing the client.
   final bool _basicAuth;
 
@@ -89,10 +95,14 @@ class Client extends http.BaseClient {
       {this.identifier,
       this.secret,
       CredentialsRefreshedCallback? onCredentialsRefreshed,
+      CredentialsRefreshingCallback? onCredentialsRefreshing,
+      CredentialsRefreshFailedCallback? onCredentialsRefreshFailed,
       bool basicAuth = true,
       http.Client? httpClient})
       : _basicAuth = basicAuth,
         _onCredentialsRefreshed = onCredentialsRefreshed,
+        _onCredentialsRefreshing = onCredentialsRefreshing,
+        _onCredentialsRefreshFailed = onCredentialsRefreshFailed,
         _httpClient = httpClient ?? http.Client() {
     if (identifier == null && secret != null) {
       throw ArgumentError('secret may not be passed without identifier.');
@@ -118,14 +128,12 @@ class Client extends http.BaseClient {
 
     List<AuthenticationChallenge> challenges;
     try {
-      challenges = AuthenticationChallenge.parseHeader(
-          response.headers['www-authenticate']!);
+      challenges = AuthenticationChallenge.parseHeader(response.headers['www-authenticate']!);
     } on FormatException {
       return response;
     }
 
-    var challenge = challenges
-        .firstWhereOrNull((challenge) => challenge.scheme == 'bearer');
+    var challenge = challenges.firstWhereOrNull((challenge) => challenge.scheme == 'bearer');
     if (challenge == null) return response;
 
     var params = challenge.parameters;
@@ -159,6 +167,7 @@ class Client extends http.BaseClient {
     // _onCredentialsRefreshed callback is only called once.
     if (_refreshingFuture == null) {
       try {
+        _onCredentialsRefreshing?.call(_credentials);
         _refreshingFuture = credentials.refresh(
           identifier: identifier,
           secret: secret,
@@ -168,6 +177,12 @@ class Client extends http.BaseClient {
         );
         _credentials = await _refreshingFuture!;
         _onCredentialsRefreshed?.call(_credentials);
+      } on AuthorizationException catch (e) {
+        _onCredentialsRefreshFailed?.call(credentials, e);
+        rethrow;
+      } on FormatException catch (e) {
+        _onCredentialsRefreshFailed?.call(credentials, e);
+        rethrow;
       } finally {
         _refreshingFuture = null;
       }